requirements.txt
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-core==2.14.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart
openai==1.3.5
langfuse==2.16.0
tenacity==8.2.3
pymongo==4.6.1
pyarrow==14.0.2
landingai
certifi==2023.7.22
requests==2.31.0
numpy
pandas
pytest==7.4.3
pytest-asyncio==0.21.1
typing-extensions==4.8.0
aiofiles==23.2.1
pymupdf

# uvicorn backend.api.main:app --reload --port 2602
# pnpm dev
